refactor: enhance field types and optimize code structure
BREAKING CHANGE: Float() renamed to Float64() for consistency

New features:
- Add comprehensive field type support (Uint8-64, Int8-32, Float32, Time, Dur, Err, RawJSON, Dict, Array)
- Extract field definitions to separate field.go file for better organization

Improvements:
- Refactor zerolog_logger.go to use addFields() helper method
- Eliminate duplicate type switch code across 8 methods (reduced ~120 lines)
- Improve error handling with native zerolog.Err() instead of string conversion

Changes:
- field.go: Add 20+ new field type functions for complete Zerolog coverage
- log.go: Remove field definitions (moved to field.go)
- zerolog_logger.go: Extract addFields() to eliminate code duplication
- examples/remote_logger: Update Float() to Float64()

+++ examples/remote_logger/main.go
@@ -316,7 +316,7 @@ func main() {
 		zllog.String("database", "mydb"))
 
 	zllog.Warn(ctx, "cache", "Cache miss rate increasing",
-		zllog.Float("rate", 0.35))
+		zllog.Float64("rate", 0.35))
 
 	// 模拟一些日志，触发批量发送
 	for i := 0; i < 15; i++ {
